Fix conv2d test
